Added Example
This is an example for a silent installation of 7-Zip (you would need to download the 7-zip installers  file into this folder).  The additional ps1 files for handling FTA and Application Capabilities registration was created by the AppV_Manage tool and are self-contained PowerShell scripts that do not depend on the module.